requirements.txt
anthropic>=0.79.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
requests>=2.31.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
arize-phoenix>=4.0.0
openinference-instrumentation-anthropic>=0.1.0
opentelemetry-sdk>=1.20.0
